test(#3928): split the header-refusal lane by CORRUPTION CLASS at the 1500-line threshold
The round-2 lite gate reported file-size FAIL: the single lane reached 1552 of
the 1500-line test limit, as a NEW file (0 -> 1552), so the campsite rule says
split rather than acknowledge (#1135).

Split by corruption class, which is what actually divides the cases — a flipped
byte makes partition_header_readiness answer Malformed / Ready-then-invalid,
while a header that RAN OUT makes it answer Incomplete, a different arm the
byte-flip cases cannot reach:

  issue_3928_corrupt_header_refusal.rs    514  — the 5 byte-flip cases
  issue_3928_truncated_header_refusal.rs  369  — the 4 truncation cases
  support/header_refusal.rs               703  — everything both need

Splitting by class (rather than by surface) keeps each lane's oracle and its
pre-fix measurements next to the cases they justify, and putting every shared
walk in one support module means neither lane can drift into its own copy of one.

No case was dropped: 5 + 4 = the same 9 substantive cases, all green, and the
truncation lane still reds 3 of 4 against the pre-round-2 src — re-verified after
the split, since a split is exactly where a case can silently stop running.
