+++ cqlite-core/tests/issue_3928_corrupt_header_refusal.rs
@@ -76,51 +76,14 @@
 //! the garbage it landed on. That is why the BIG case asserts the ABSENCE of the
 //! resync WARN as well as the refusal — see its own comment.
 //!
-//! # Fix round 1 — a TRUNCATED header, and two more swallows
+//! # The SIBLING lane — a header that RAN OUT (`Incomplete`)
 //!
 //! A flipped byte and a truncated file are different corruption classes and
-//! reach different classifier arms, so review found two more swallows that the
-//! byte-flip cases above cannot see. Both were measured on a prefix of this same
-//! Cassandra-written section, cut inside the LAST partition's header:
-//!
-//! * **the two stitched walks DISAGREED.** With one header byte surviving, the
-//!   block walk answered `Ok` with **99** of 100 partition keys while the
-//!   cell-metadata walk answered `Err("Unexpected end at partition key
-//!   length")` — on the same bytes. `data_access/mod.rs:249` and `:288` hand
-//!   both walks the SAME stitched `Complete` buffer, so that is `SELECT *`
-//!   answering `Ok` and `SELECT *, WRITETIME(c)` answering `Err` over one file.
-//! * **the sliding drivers reported a truncated header as CLEAN COMPLETION.**
-//!   `PartitionHeaderReadiness::Incomplete` at the final chunk became
-//!   `DriverHeader::Done` for every cause, so a header declaring a 16-byte key
-//!   with only 10 bytes present answered `Done` and dropped the partition.
-//!
-//! Both are pinned by `both_stitched_walks_agree_and_refuse_a_truncated_final_header`
-//! and `the_sliding_driver_refuses_any_non_empty_truncated_header_at_the_final_chunk`.
-//!
-//! # Fix round 2 — one predicate, and the two edges the round-1 one got wrong
-//!
-//! Review found the round-1 predicate wrong at BOTH its edges, which is the
-//! third consecutive finding in one place — so the mechanism was replaced rather
-//! than given a fourth condition. Tolerance is now ONE piece of state,
-//! `HeaderTolerance` (in `buffer_extent.rs`, beside the contract it completes):
-//! *can a byte still arrive, or is this walk's progress no longer attributable?*
-//! Both header arms consult it and the byte count decides nothing.
-//!
-//! * **C1** — a ONE-BYTE final chunk was returned as clean completion. That byte
-//!   can be the surviving first byte of a truncated key length, so both drivers
-//!   dropped the final partition; and `block_emit` already REFUSED the identical
-//!   bytes, so the DRIVER path and the BLOCK-EMIT path disagreed about one file —
-//!   a different pair from the block-vs-metadata one round 1 closed. Measured
-//!   pre-fix: `Done after emitting 0 row(s) over 1 byte(s) of a header that
-//!   declares a 16-byte key`, and three-way `driver: Ok, block-emit walk: Err,
-//!   cell-metadata walk: Err`. Pinned by
-//!   `the_driver_and_both_block_walks_agree_on_a_truncated_final_header`.
-//! * **C2** — the round-1 flag disabled refusal for the WHOLE call, so a
-//!   `Complete` buffer with a `row_body_window` byte-resynced past the malformed
-//!   INITIAL header, before the window had introduced any uncertainty. Measured
-//!   pre-fix: `Skipping malformed partition header at offset 0 … key length of
-//!   0 … (partition=0)`. Pinned by
-//!   `a_bounded_walk_still_refuses_a_malformed_initial_header`.
+//! reach different classifier arms, so the cases here CANNOT see a truncation
+//! swallow — review found two, plus a driver-vs-block-emit divergence. Those
+//! live in `issue_3928_truncated_header_refusal.rs`, with their own pre-fix
+//! measurements; the scaffolding both lanes share is
+//! `support/header_refusal.rs`.
 //!
 //! # DECLARED RESIDUAL — the partition-key LENGTH model (#3999)
 //!
@@ -135,7 +98,6 @@
 //! 256-byte-key fixture the corpus does not have — **#3999**. The refusal stays
 //! (loud beats silent, which is this issue's whole subject) and names the issue.
 //! No corpus table has a key that long, which is why the negative control below
-//! is green.
 #![cfg(all(
     feature = "state_machine",
     feature = "cli-helpers",
@@ -149,14 +111,10 @@
 ))]
 
 use std::collections::BTreeSet;
-use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
-use cqlite_core::ingestion::{ingest, IngestionConfig};
 use cqlite_core::platform::Platform;
-use cqlite_core::schema::TableSchema;
 use cqlite_core::storage::sstable::SSTableReader;
-use cqlite_core::Database;
 
 #[path = "support/datasets_root.rs"]
 mod datasets_root;
@@ -164,378 +122,12 @@ mod datasets_root;
 mod fixture;
 #[path = "support/multiset.rs"]
 mod multiset;
+#[path = "support/header_refusal.rs"]
+mod support;
 
-use fixture::{FixtureSpec, BIG_COMPOSITE_HEADER, BTI_MULTICLUSTERING_HEADER};
+use fixture::{BIG_COMPOSITE_HEADER, BTI_MULTICLUSTERING_HEADER};
+use support::*;
 
-/// The fixture's GENERATION directory, resolved per TABLE (#3220) so a root that
-/// holds the keyspace but not this table cannot silently win the selection. The
-/// two specs live in DIFFERENT roots on a fleet box — the BTI one is
-/// git-committed in the checkout, the BIG one is fetched-corpus-only — so
-/// evidence, never a preference order, decides (#3104). "Not found" is a loud
-/// named panic, never a skip.
-fn fixture_dir(spec: &FixtureSpec) -> PathBuf {
-    datasets_root::resolve_table_generation_dir(spec.keyspace, spec.table).unwrap_or_else(|why| {
-        panic!(
-            "fixture {}.{} has no usable generation directory: {why}",
-            spec.keyspace, spec.table
-        )
-    })
-}
-
-fn schema_file(spec: &FixtureSpec) -> PathBuf {
-    datasets_root::schema_path(spec.schema_file).expect("committed CQL schema (#3148)")
-}
-
-fn table_schema(spec: &FixtureSpec) -> TableSchema {
-    let cql = std::fs::read_to_string(schema_file(spec)).expect("read schema");
-    let start = cql
-        .find(&format!("CREATE TABLE IF NOT EXISTS {}", spec.table))
-        .expect("CREATE TABLE statement");
-    let end = start + cql[start..].find(';').expect("statement terminator") + 1;
-    let mut t = cqlite_core::schema::cql_parser::parse_cql_schema(&cql[start..end])
-        .expect("parse CREATE TABLE");
-    t.keyspace = spec.keyspace.to_string();
-    t
-}
-
-async fn open_db(spec: &FixtureSpec, data_dir: PathBuf) -> Database {
-    ingest(IngestionConfig {
-        schema_paths: vec![schema_file(spec)],
-        data_dir,
-        version_hint: None,
-        core_config: cqlite_core::Config::default(),
-        table_directory_filter: Some(format!("/{}/", spec.keyspace)),
-    })
-    .await
-    .expect("ingest")
-    .database
-}
-
-async fn open_reader(dir: &Path) -> SSTableReader {
-    let config = cqlite_core::Config::default();
-    let platform = Arc::new(Platform::new(&config).await.expect("platform"));
-    SSTableReader::open(&fixture::comp_file(dir, "-Data.db"), &config, platform)
-        .await
-        .expect("open SSTableReader")
-}
-
-/// What one read surface answered on one leg of a staged pair.
-///
-/// `keys` is `Err(message)` when the surface REFUSED and `Ok(keys)` when it
-/// answered `Ok` — carrying the partition key of every emitted element, in emit
-/// order, so a DUPLICATE is visible. A set would hide the resync's re-emission
-/// of a partition it had already emitted, which is one of the ways the
-/// #3782/#3928 shape makes the count go UP while data is lost.
-///
-/// The refusal's MESSAGE is retained rather than discarded: a control leg that
-/// refuses a PRISTINE fixture is a broken lane, and "it refused" without saying
-/// why sends the next reader back to the debugger.
-struct Outcome {
-    name: &'static str,
-    keys: Result<Vec<Vec<u8>>, String>,
-}
-
-impl Outcome {
-    fn describe(&self, control: &std::collections::BTreeMap<Vec<u8>, usize>) -> String {
-        match &self.keys {
-            Err(why) => format!("{}: REFUSED ({why})", self.name),
-            Ok(keys) => {
-                let got = multiset::multiset(keys.iter().cloned());
-                let lost = multiset::deficit(&got, control);
-                let fabricated = multiset::surplus(&got, control);
-                format!(
-                    "{}: Ok with {} emitted key occurrence(s); {} LOST [{}], {} FABRICATED [{}]",
-                    self.name,
-                    keys.len(),
-                    lost.iter().map(|(_, n)| n).sum::<usize>(),
-                    multiset::describe(&lost),
-                    fabricated.iter().map(|(_, n)| n).sum::<usize>(),
-                    multiset::describe(&fabricated),
-                )
-            }
-        }
-    }
-}
-
-/// Render a refusal for a diagnostic.
-fn why(e: cqlite_core::Error) -> String {
-    e.to_string()
-}
-
-/// A WARN-level subscriber writing into `sink`, for `tracing::subscriber::set_default`.
-fn warn_subscriber(sink: &LogSink) -> impl tracing::Subscriber + Send + Sync {
-    tracing_subscriber::fmt()
-        .with_writer(sink.clone())
-        .with_max_level(tracing::Level::WARN)
-        .with_ansi(false)
-        .finish()
-}
-
-/// A control-leg surface's own partition-key multiset. An empty multiset for a
-/// surface that REFUSED the control leg, which `assert_control_is_healthy`
-/// already panics on — this only keeps the rendering total.
-fn control_multiset(c: &Outcome) -> std::collections::BTreeMap<Vec<u8>, usize> {
-    c.keys
-        .as_ref()
-        .map(|k| multiset::multiset(k.iter().cloned()))
-        .unwrap_or_default()
-}
-
-/// Every partition-key-bearing read surface, evaluated over the generation in
-/// `dir`.
-///
-/// The set deliberately spans BOTH propagation mechanisms this issue touches,
-/// because they are different code with different discriminators and a fix to
-/// one says nothing about the other:
-///
-/// * `distinct_partition_keys` / `partition_verify_scan` / `get_all_entries` /
-///   `iterate_all_partitions` reach `parse_block_emit*`, whose discriminator is
-///   the explicit `BufferExtent` the caller states;
-/// * `iterate_all_partitions_for_compaction` reaches `drive_partition_sliding`
-///   and `stream_all_partitions_for_compaction` reaches
-///   `stream_partition_body_incremental`, whose discriminator is the driver's own
-///   `at_final_chunk`.
-async fn observe(dir: &Path, schema: &TableSchema) -> Vec<Outcome> {
-    let reader = open_reader(dir).await;
-    let mut out: Vec<Outcome> = Vec::new();
-
-    out.push(Outcome {
-        name: "distinct_partition_keys",
-        keys: reader.distinct_partition_keys().await.map_err(why),
-    });
-    out.push(Outcome {
-        name: "partition_verify_scan",
-        keys: reader
-            .partition_verify_scan()
-            .await
-            .map_err(why)
-            .map(|rows| rows.into_iter().map(|(k, _ldt)| k).collect()),
-    });
-    out.push(Outcome {
-        name: "get_all_entries",
-        keys: reader.get_all_entries().await.map_err(why).map(|rows| {
-            rows.into_iter()
-                .map(|(_t, k, _r)| k.as_bytes().to_vec())
-                .collect()
-        }),
-    });
-    out.push(Outcome {
-        name: "iterate_all_partitions",
-        keys: reader
-            .iterate_all_partitions()
-            .await
-            .map_err(why)
-            .map(|rows| {
-                rows.into_iter()
-                    .map(|(k, _r)| k.as_bytes().to_vec())
-                    .collect()
-            }),
-    });
-    out.push(Outcome {
-        name: "iterate_all_partitions_for_compaction",
-        keys: reader
-            .iterate_all_partitions_for_compaction(Some(schema))
-            .await
-            .map_err(why)
-            .map(|rows| {
-                rows.into_iter()
-                    .map(|r| r.key.as_bytes().to_vec())
-                    .collect()
-            }),
-    });
-
-    let cancel = cqlite_core::storage::scan_cancel::ScanCancel::new();
-    let mut streamed: Vec<Vec<u8>> = Vec::new();
-    let streamed = reader
-        .stream_all_partitions_for_compaction(Some(schema), &cancel, |row| {
-            streamed.push(row.key.as_bytes().to_vec());
-            Ok(std::ops::ControlFlow::Continue(()))
-        })
-        .await
-        .map_err(why)
-        // The rows emitted BEFORE a refusal are deliberately discarded: a
-        // surface that refused emitted no answer, and counting its partial
-        // emission as an answer would let a refusal be scored for fabrication.
-        .map(|()| streamed);
-    out.push(Outcome {
-        name: "stream_all_partitions_for_compaction",
-        keys: streamed,
-    });
-
-    out
-}
-
-/// The control leg's partition-key multiset, plus the proof it is non-empty
-/// (0-rows-when-present is a failure, never a pass).
-async fn control_keys(
-    dir: &Path,
-    spec: &FixtureSpec,
-) -> std::collections::BTreeMap<Vec<u8>, usize> {
-    let reader = open_reader(dir).await;
-    let keys = reader.distinct_partition_keys().await.unwrap_or_else(|e| {
-        panic!(
-            "the pristine {}.{} must enumerate its partition keys: {e}",
-            spec.keyspace, spec.table
-        )
-    });
-    assert!(
-        !keys.is_empty(),
-        "0-rows-when-present: the pristine {}.{} control enumerated no partition keys",
-        spec.keyspace,
-        spec.table
-    );
-    multiset::multiset(keys)
-}
-
-/// Assert that EVERY surface answered on the well-formed control leg, and agreed
-/// on WHICH partitions the fixture holds — so "the mutated leg refused" can
-/// never be a surface that refuses on healthy data too, and the per-surface
-/// controls the fabrication check compares against are themselves sound.
-///
-/// The comparison is over the DISTINCT key set, not the multiset: these surfaces
-/// have deliberately different granularities (`distinct_partition_keys` emits
-/// one entry per PARTITION, `get_all_entries` one per ROW), so their
-/// multiplicities legitimately differ — measured on the BTI fixture, 3 vs 468.
-/// The per-surface multiset is still the oracle for FABRICATION below; it is
-/// just compared against the SAME surface on the pristine leg.
-fn assert_control_is_healthy(
-    observed: &[Outcome],
-    control_partitions: &std::collections::BTreeMap<Vec<u8>, usize>,
-    spec: &FixtureSpec,
-) {
-    let expected: BTreeSet<&Vec<u8>> = control_partitions.keys().collect();
-    for o in observed {
-        let keys = o.keys.as_ref().unwrap_or_else(|why| {
-            panic!(
-                "control leg: {}.{} surface `{}` REFUSED a PRISTINE Cassandra fixture ({why}) \
-                 — the mutated-leg expectations below would be meaningless",
-                spec.keyspace, spec.table, o.name
-            )
-        });
-        assert!(
-            !keys.is_empty(),
-            "0-rows-when-present: {}.{} surface `{}` answered Ok with NOTHING on a pristine \
-             fixture",
-            spec.keyspace,
-            spec.table,
-            o.name
-        );
-        let got: BTreeSet<&Vec<u8>> = keys.iter().collect();
-        assert_eq!(
-            got.len(),
-            expected.len(),
-            "control leg: {}.{} surface `{}` sees {} distinct partition key(s) where \
-             `distinct_partition_keys` sees {} — the surfaces disagree about the PRISTINE \
-             fixture, so neither can be an oracle for the mutated one",
-            spec.keyspace,
-            spec.table,
-            o.name,
-            got.len(),
-            expected.len()
-        );
-        assert!(
-            got == expected,
-            "control leg: {}.{} surface `{}` reports a DIFFERENT set of partition keys than \
-             `distinct_partition_keys` on the PRISTINE fixture",
-            spec.keyspace,
-            spec.table,
-            o.name
-        );
-    }
-}
-
-/// The surfaces of the mutated leg that still answered `Ok`, described against
-/// the control leg's SAME-SURFACE result.
-///
-/// Pairing surface-with-itself is what makes a fabrication verdict meaningful:
-/// these surfaces emit at different granularities, so a cross-surface multiset
-/// comparison would report hundreds of spurious "surplus" occurrences (measured:
-/// 465 on the BTI fixture, from `get_all_entries`' per-ROW emit against a
-/// per-PARTITION control).
-fn tolerating(control: &[Outcome], mutated: &[Outcome]) -> Vec<String> {
-    assert_eq!(
-        control.len(),
-        mutated.len(),
-        "the two legs must be observed through the same surface list"
-    );
-    control
-        .iter()
-        .zip(mutated.iter())
-        .filter(|(_, m)| m.keys.is_ok())
-        .map(|(c, m)| m.describe(&control_multiset(c)))
-        .collect()
-}
-
-/// The surfaces of the mutated leg that FABRICATED — answered `Ok` carrying a
-/// partition-key occurrence the SAME surface did not produce on the pristine
-/// fixture.
-fn fabricating(control: &[Outcome], mutated: &[Outcome]) -> Vec<String> {
-    assert_eq!(
-        control.len(),
-        mutated.len(),
-        "the two legs must be observed through the same surface list"
-    );
-    control
-        .iter()
-        .zip(mutated.iter())
-        .filter_map(|(c, m)| {
-            let keys = m.keys.as_ref().ok()?;
-            let ctl = control_multiset(c);
-            let got = multiset::multiset(keys.iter().cloned());
-            if multiset::surplus(&got, &ctl).is_empty() {
-                None
-            } else {
-                Some(m.describe(&ctl))
-            }
-        })
-        .collect()
-}
-
-/// Capture WARN-and-above tracing output into a shared buffer, so a case can
-/// assert what a code path did NOT log.
-#[derive(Clone, Default)]
-struct LogSink(Arc<std::sync::Mutex<Vec<u8>>>);
-impl std::io::Write for LogSink {
-    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-        self.0
-            .lock()
-            .expect("log sink mutex")
-            .extend_from_slice(buf);
-        Ok(buf.len())
-    }
-    fn flush(&mut self) -> std::io::Result<()> {
-        Ok(())
-    }
-}
-impl<'a> tracing_subscriber::fmt::MakeWriter<'a> for LogSink {
-    type Writer = LogSink;
-    fn make_writer(&'a self) -> LogSink {
-        self.clone()
-    }
-}
-impl LogSink {
-    fn text(&self) -> String {
-        String::from_utf8_lossy(&self.0.lock().expect("log sink mutex").clone()).to_string()
-    }
-}
-
-/// The two WARNs the pre-#3928 block-emit header arm emitted immediately before
-/// `offset += 1`, verbatim from
-/// `block_emit_windowed/partition_header_arm.rs`. Since the fix they are
-/// reachable ONLY on a `BufferExtent::Window`, so their presence on a
-/// full-extent walk IS the resync.
-const RESYNC_WARNS: &[&str] = &[
-    "Skipping malformed partition header",
-    "Failed to parse partition header",
-];
-
-/// AC1 — on a PROVEN-COMPLETE buffer a malformed partition header REFUSES.
-///
-/// The BIG (`nb`) fixture is a SINGLE compressed chunk, so every surface below
-/// reaches its parse with the complete extent stated (`BufferExtent::Complete`)
-/// or at the sliding driver's final chunk (`at_final_chunk == true`) — the two
-/// authoritative "no further bytes can arrive" signals. There is therefore no
-/// straddle reading of the failure available: the header is corrupt, and the
 /// partition is lost unless the error is reported.
 #[tokio::test]
 async fn every_proven_complete_surface_refuses_a_malformed_partition_header() {
@@ -919,634 +511,4 @@ async fn corpus_wide_well_formed_tables_still_compact_without_refusal() {
          partition",
         scanned_tables.len()
     );
-}
-
-/// Committed/required corpus subjects the corpus case MUST scan — fail-closed
-/// unconditionally, never a skip (#3220). Same pair, and for the same reason, as
-/// `issue_3782_corrupt_row_refusal.rs`: neither candidate root is a superset of
-/// the other, so the scan takes their union and resolves each table by evidence.
-const MUST_RUN: &[(&str, &str)] = &[
-    ("test_da", "multiclustering_table"),
-    ("test_basic", "composite_key_table"),
-];
-
-/// Every `(keyspace, table)` identity that carries a real `*-Data.db` under ANY
-/// candidate root — the UNION, deduplicated by identity. A `break` on the first
-/// non-empty root is a PREFERENCE ORDERING and misses the checkout-only tables
-/// (#3220).
-fn corpus_table_identities() -> BTreeSet<(String, String)> {
-    let mut ids: BTreeSet<(String, String)> = BTreeSet::new();
-    for root in datasets_root::sstables_root_candidates() {
-        let Ok(keyspaces) = std::fs::read_dir(&root) else {
-            continue;
-        };
-        for ks in keyspaces.flatten() {
-            if !ks.path().is_dir() {
-                continue;
-            }
-            let keyspace = ks.file_name().to_string_lossy().to_string();
-            let Ok(tables) = std::fs::read_dir(ks.path()) else {
-                continue;
-            };
-            for table in tables.flatten() {
-                if !table.path().is_dir() {
-                    continue;
-                }
-                let dir_name = table.file_name().to_string_lossy().to_string();
-                // `<table>-<generation uuid>`; a CQL table name cannot contain
-                // `-`, so the last separator is the generation boundary.
-                let Some((name, _generation)) = dir_name.rsplit_once('-') else {
-                    continue;
-                };
-                if datasets_root::table_has_data(&root, &keyspace, name) {
-                    ids.insert((keyspace.clone(), name.to_string()));
-                }
-            }
-        }
-    }
-    ids
-}
-
-/// Every `*-Data.db` of `<keyspace>.<table>` under the root that EVIDENCE picks
-/// for that table (`sstables_root_for_table`, the sanctioned per-table resolver),
-/// across all of that table's generation directories.
-fn data_files_for_table(keyspace: &str, table: &str) -> Vec<PathBuf> {
-    let Some(root) = datasets_root::sstables_root_for_table(keyspace, table) else {
-        return Vec::new();
-    };
-    let mut out: Vec<PathBuf> = Vec::new();
-    for dir in datasets_root::table_generation_dirs(&root, keyspace, table) {
-        if let Ok(rd) = std::fs::read_dir(&dir) {
-            for f in rd.flatten() {
-                let p = f.path();
-                if p.to_string_lossy().ends_with("-Data.db") {
-                    out.push(p);
-                }
-            }
-        }
-    }
-    out.sort();
-    out
-}
-
-/// Assert the refusal's KIND, not a message substring.
-///
-/// A message check stays green through a refactor that re-wraps the decode error
-/// in a different variant while forwarding the text, which is the no-heuristics
-/// shape: it reads bytes of a rendered string instead of the authoritative
-/// discriminant.
-fn assert_corruption_kind(e: &cqlite_core::Error, surface: &str) {
-    assert!(
-        matches!(e, cqlite_core::Error::Corruption(_)),
-        "{surface}: the refusal must carry the header decode error's own kind \
-         (Error::Corruption), not a re-wrapped generic; got {e:?}"
-    );
-}
-
-// ===========================================================================
-// Fix round 1 — a header that RAN OUT, which is a different corruption class
-// from a header with a flipped byte and reaches different arms.
-//
-// A flipped byte makes `partition_header_readiness` answer `Malformed` (a
-// zero/over-long declared key length) or `Ready` (present but structurally
-// invalid). A `Data.db` that ENDS inside a header makes it answer
-// `Incomplete` — and at the sliding drivers' FINAL chunk that answer was
-// `DriverHeader::Done`, which both drivers read as CLEAN COMPLETION. A header
-// with a valid nonzero key length whose key or `DeletionTime` was truncated
-// away is well over two bytes and is not a tail; reporting it as completion is
-// the same swallow, one classifier arm over.
-//
-// The bytes are a PREFIX of the fixture's own stitched decompressed data
-// section — exactly what a file truncated by a crashed writer or a short read
-// presents — and the cut offset comes from `Index.db`, cross-checked against
-// `Data.db` (see `last_partition_header_offset`). No file is rewritten, so
-// there is no re-compression or CRC arithmetic to get wrong.
-// ===========================================================================
-
-/// Where a truncated `Data.db` ends inside the LAST partition's header.
-///
-/// The three variants are the three distinct answers
-/// `partition_header_readiness` gives for a header that ran out, and each is
-/// decided by a DIFFERENT branch of that classifier — so covering one says
-/// nothing about the others.
-#[derive(Clone, Copy, Debug)]
-enum HeaderCut {
-    /// ONE byte of the header survives: fewer than the two any header needs.
-    /// The classifier's `data.len() < 2` branch. This is the tail the block-emit
-    /// walk deliberately tolerates, and the one the drivers report as `Done`.
-    OneByte,
-    /// The 2-byte key length survives and the KEY is cut in half. The
-    /// classifier's `data.get(deletion_offset) == None` branch.
-    InsideKey,
-    /// Key length and key survive; the partition-level `DeletionTime` is cut
-    /// (`nb` needs 12 bytes and 4 are present). The classifier's
-    /// `deletion_offset + deletion_time_min > data.len()` branch.
-    InsideDeletionTime,
-}
-
-/// The pristine stitched section, the last partition header's offset, and the
-/// byte count that leaves that header cut as `cut` describes.
-fn truncation(dir: &Path, cut: HeaderCut) -> (Vec<u8>, usize, usize) {
-    let dec = fixture::stitched_data_section(dir);
-    let hdr = fixture::last_partition_header_offset(dir, &dec);
-    // `writeWithShortLength`: 2-byte big-endian key length
-    // (`SortedTablePartitionWriter.start`, cassandra-5.0.8). Cross-checked
-    // against Index.db by the helper above, so this is a measured length.
-    let key_len = usize::from(u16::from_be_bytes([dec[hdr], dec[hdr + 1]]));
-    assert!(
-        key_len >= 4,
-        "this fixture's partition key is {key_len} byte(s); the InsideKey cut needs at \
-         least 4 so that some key bytes survive AND some are cut"
-    );
-    let keep = match cut {
-        HeaderCut::OneByte => hdr + 1,
-        HeaderCut::InsideKey => hdr + 2 + key_len / 2,
-        // `nb`'s DeletionTime is a fixed 12 bytes (`DeletionTime.LegacySerializer`:
-        // 4-byte localDeletionTime + 8-byte markedForDeleteAt), so 4 present is a
-        // genuine mid-field cut.
-        HeaderCut::InsideDeletionTime => hdr + 2 + key_len + 4,
-    };
-    assert!(
-        keep > hdr && keep < dec.len(),
-        "the {cut:?} cut must land strictly inside the last header: keep={keep}, \
-         header at {hdr}, section {} byte(s)",
-        dec.len()
-    );
-    (dec, hdr, keep)
-}
-
-/// A parser configured the way the stitched read paths configure theirs, minus
-/// the Statistics.db timestamp bases (which affect decoded VALUES, never
-/// framing). The control assertions in each case below require this parser to
-/// decode the PRISTINE section completely, so its adequacy for these cases is
-/// measured rather than assumed.
-fn framing_parser(
-    spec: &FixtureSpec,
-) -> cqlite_core::storage::sstable::reader::V5CompressedLegacyParser {
-    cqlite_core::storage::sstable::reader::V5CompressedLegacyParser::new(
-        spec.keyspace.to_string(),
-        spec.table.to_string(),
-        0,
-        0,
-        None,
-    )
-}
-
-/// Partition keys the block-emit walk (`parse_block` → `parse_block_emit_windowed`
-/// with NO row-body window, the route `stitch_and_parse_all_chunks` takes) emits
-/// over `buf`, or the refusal.
-fn block_walk(
-    spec: &FixtureSpec,
-    schema: &TableSchema,
-    reader: &SSTableReader,
-    buf: &[u8],
-) -> Result<Vec<Vec<u8>>, String> {
-    use cqlite_core::storage::sstable::reader::BufferExtent;
-    let mut keys: Vec<Vec<u8>> = Vec::new();
-    framing_parser(spec)
-        .parse_block_emit_windowed(
-            buf,
-            BufferExtent::Complete,
-            Some(schema),
-            reader,
-            None,
-            |(_t, k, _r)| {
-                keys.push(k.as_bytes().to_vec());
-                Ok(std::ops::ControlFlow::Continue(()))
-            },
-        )
-        .map_err(why)
-        .map(|()| keys)
-}
-
-/// The same buffer through the CELL-METADATA walk — `parse_block_with_cell_metadata`,
-/// the route `stitch_and_parse_all_chunks_with_metadata` takes when a query
-/// carries `WRITETIME(col)`/`TTL(col)` (`ProjectionFlags::include_cell_metadata`).
-fn metadata_walk(
-    spec: &FixtureSpec,
-    schema: &TableSchema,
-    reader: &SSTableReader,
-    buf: &[u8],
-) -> Result<Vec<Vec<u8>>, String> {
-    use cqlite_core::storage::sstable::reader::BufferExtent;
-    framing_parser(spec)
-        .parse_block_with_cell_metadata(buf, BufferExtent::Complete, Some(schema), reader)
-        .map_err(why)
-        .map(|rows| {
-            rows.into_iter()
-                .map(|(_t, k, _r, _m)| k.as_bytes().to_vec())
-                .collect()
-        })
-}
-
-/// The SLIDING driver, at its FINAL chunk — `parse_one_partition_for_compaction`,
-/// the public entry to `drive_partition_sliding`, which is also what
-/// `stream_partition_body_incremental` shares its header arm with.
-fn driver_at_final_chunk(
-    spec: &FixtureSpec,
-    schema: &TableSchema,
-    reader: &SSTableReader,
-    buf: &[u8],
-) -> Result<String, String> {
-    let mut emitted = 0usize;
-    framing_parser(spec)
-        .parse_one_partition_for_compaction(buf, Some(schema), reader, true, &mut |_row| {
-            emitted += 1;
-            Ok(std::ops::ControlFlow::Continue(()))
-        })
-        .map_err(why)
-        .map(|step| format!("{step:?} after emitting {emitted} row(s)"))
-}
-
-/// B1(b) + the `SELECT *` vs `WRITETIME(col)` divergence — the two stitched
-/// walks must AGREE, and on an unbounded proven-complete buffer both must
-/// REFUSE.
-///
-/// `stitch_and_parse_all_chunks` (`data_access/mod.rs:249`) and
-/// `stitch_and_parse_all_chunks_with_metadata` (`:288`) hand the SAME
-/// `stitch_all_chunks` buffer to their parses with the SAME
-/// `BufferExtent::Complete`; they are the plain and
-/// `WRITETIME`/`TTL`-projection variants of ONE query. So before this round a
-/// `Data.db` truncated to one surviving header byte answered `SELECT *` with
-/// `Ok` (last partition silently dropped, via the block walk's tail carve-out)
-/// and `SELECT *, WRITETIME(c)` with `Err` (the metadata walk has no such
-/// carve-out) — on the same file.
-///
-/// This case asserts the property at the parse both routes are handed, which is
-/// where the divergence lives; it does not drive the two SQL statements, because
-/// that would need a rewritten `CompressionInfo.db` and CRC to truncate the file
-/// on disk, and the buffer is the same object either way.
-#[tokio::test]
-async fn both_stitched_walks_agree_and_refuse_a_truncated_final_header() {
-    let spec = &BIG_COMPOSITE_HEADER;
-    let dir = fixture_dir(spec);
-    let schema = table_schema(spec);
-    let reader = open_reader(&dir).await;
-
-    for cut in [
-        HeaderCut::OneByte,
-        HeaderCut::InsideKey,
-        HeaderCut::InsideDeletionTime,
-    ] {
-        let (dec, hdr, keep) = truncation(&dir, cut);
-
-        // CONTROL: the untruncated section decodes completely on BOTH walks, and
-        // they agree. This is what makes a refusal below attributable to the
-        // truncation rather than to this parser's configuration.
-        let control_block = block_walk(spec, &schema, &reader, &dec).unwrap_or_else(|e| {
-            panic!("{cut:?}: the PRISTINE section must decode on the block walk: {e}")
-        });
-        let control_meta = metadata_walk(spec, &schema, &reader, &dec).unwrap_or_else(|e| {
-            panic!("{cut:?}: the PRISTINE section must decode on the metadata walk: {e}")
-        });
-        assert!(
-            !control_block.is_empty(),
-            "0-rows-when-present: {cut:?} control block walk emitted nothing"
-        );
-        assert_eq!(
-            multiset::multiset(control_block.iter().cloned()),
-            multiset::multiset(control_meta.iter().cloned()),
-            "{cut:?}: the two stitched walks must agree on the PRISTINE section"
-        );
-
-        let got_block = block_walk(spec, &schema, &reader, &dec[..keep]);
-        let got_meta = metadata_walk(spec, &schema, &reader, &dec[..keep]);
-
-        // The DIVERGENCE assertion: whatever the answer is, one query variant may
-        // not disagree with the other about the same bytes.
-        assert_eq!(
-            got_block.is_err(),
-            got_meta.is_err(),
-            "{cut:?} (truncated to {keep} of {} bytes, last header at {hdr}): the two \
-             stitched walks DISAGREE — block walk {} / metadata walk {}. These are \
-             `SELECT *` and `SELECT *, WRITETIME(col)` over one file.",
-            dec.len(),
-            got_block
-                .as_ref()
-                .map_or_else(|e| format!("Err({e})"), |k| format!("Ok({} keys)", k.len())),
-            got_meta
-                .as_ref()
-                .map_or_else(|e| format!("Err({e})"), |k| format!("Ok({} keys)", k.len())),
-        );
-
-        // And on an UNBOUNDED, proven-complete buffer both must refuse: the
-        // partition Cassandra wrote there is gone, so answering `Ok` reports a
-        // table with one fewer partition than the file was written with.
-        if let Ok(keys) = &got_block {
-            let got = multiset::multiset(keys.iter().cloned());
-            let control = multiset::multiset(control_block.iter().cloned());
-            let lost = multiset::deficit(&got, &control);
-            panic!(
-                "{cut:?} (truncated to {keep} of {} bytes): the block walk answered Ok with \
-                 {} key occurrence(s) against a control of {} — {} LOST [{}]. A truncated \
-                 header on a proven-complete buffer is DATA LOSS and must be refused.",
-                dec.len(),
-                keys.len(),
-                control_block.len(),
-                lost.iter().map(|(_, n)| n).sum::<usize>(),
-                multiset::describe(&lost)
-            );
-        }
-    }
-}
-
-/// B1(a)/C1 — the sliding drivers' `Incomplete`-at-the-final-chunk answer.
-///
-/// `DriverHeader::Done` is read by BOTH drivers as clean completion
-/// (`ParseStep::Done` / `PartitionStreamStep::AllDone`), so it is truthful only
-/// where there is genuinely nothing there. At the final chunk a NON-EMPTY
-/// incomplete header is a header Cassandra wrote whose bytes were truncated
-/// away — including the one-byte case, because that byte can be the surviving
-/// first byte of a partition key length.
-///
-/// Round 2 (C1) made the one-byte case fatal too. It had been kept tolerant as a
-/// both-directions guard, which was guarding the wrong thing: the proof that
-/// this case cannot pass by refusing everything now comes from a leg that is
-/// LEGITIMATELY tolerant — the same cuts at `at_final_chunk == false`, where
-/// more bytes can still arrive and `NeedMore` is the correct answer.
-///
-/// Evidence that no caller legitimately leaves one byte at the final chunk,
-/// since making it fatal rests on that: (1) both drivers early-return on an
-/// EMPTY buffer (`partition_driver.rs`, `compaction_stream.rs`), so the arm sees
-/// at least one byte by construction; (2) every `Emitted(consumed)` /
-/// `PartitionDone(consumed)` counts the structural terminator, and the callers
-/// advance by exactly that (`scan_stream_windowed.rs`'s `window.consume(take)`,
-/// `drain_compaction_window`), so a well-formed walk leaves ZERO bytes, never
-/// one; (3) measured — the AC3 counters recorded ZERO arrivals at this arm at
-/// the final chunk across 126 tables / 148 SSTables on four surfaces, two of
-/// which drive these drivers.
-#[tokio::test]
-async fn the_sliding_driver_refuses_any_non_empty_truncated_header_at_the_final_chunk() {
-    let spec = &BIG_COMPOSITE_HEADER;
-    let dir = fixture_dir(spec);
-    let schema = table_schema(spec);
-    let reader = open_reader(&dir).await;
-
-    // CONTROL: the last partition, whole, drives to an Emitted step.
-    let (dec, hdr, _) = truncation(&dir, HeaderCut::OneByte);
-    let whole = driver_at_final_chunk(spec, &schema, &reader, &dec[hdr..]).unwrap_or_else(|e| {
-        panic!("the PRISTINE last partition must drive cleanly at the final chunk: {e}")
-    });
-    assert!(
-        whole.starts_with("Emitted"),
-        "control: the whole last partition must report Emitted(consumed), got {whole}"
-    );
-
-    for cut in [
-        HeaderCut::OneByte,
-        HeaderCut::InsideKey,
-        HeaderCut::InsideDeletionTime,
-    ] {
-        let (dec, hdr, keep) = truncation(&dir, cut);
-        let declared = u16::from_be_bytes([dec[hdr], dec[hdr + 1]]);
-
-        // The LEGITIMATELY-TOLERANT leg: the same bytes mid-stream, where a
-        // header may still be completed by the next chunk. `NeedMore` is the
-        // correct answer and this case must not turn it into an error — that is
-        // what stops it passing by refusing everything.
-        let midstream =
-            driver_mid_stream(spec, &schema, &reader, &dec[hdr..keep]).unwrap_or_else(|e| {
-                panic!(
-                    "{cut:?}: mid-stream a truncated header must ask for more bytes, not \
-                     refuse — more bytes can still arrive there: {e}"
-                )
-            });
-        assert!(
-            midstream.starts_with("NeedMore"),
-            "{cut:?}: mid-stream must report NeedMore, got {midstream}"
-        );
-
-        // At the final chunk every one of them is DATA LOSS.
-        let got = driver_at_final_chunk(spec, &schema, &reader, &dec[hdr..keep]);
-        assert!(
-            got.is_err(),
-            "C1 {cut:?}: a NON-EMPTY truncated header must be REFUSED at the final chunk, \
-             not reported as completion. The driver answered {} over {} byte(s) of a header \
-             that declares a {declared}-byte key — the partition Cassandra wrote there is \
-             gone, and both drivers read that answer as a clean end of walk.",
-            got.as_ref().map_or_else(|e| e.clone(), |s| s.clone()),
-            keep - hdr
-        );
-    }
-}
-
-/// C1's second half — the DRIVER path and the two BLOCK-EMIT walks must agree.
-///
-/// `both_stitched_walks_agree_and_refuse_a_truncated_final_header` pins the
-/// block-vs-metadata pair, so it stays green while a DRIVER-vs-block-emit
-/// divergence survives — and one did: with one header byte surviving, the
-/// drivers reported clean completion while `block_emit` refused the identical
-/// bytes. Three surfaces, ONE buffer, one verdict.
-#[tokio::test]
-async fn the_driver_and_both_block_walks_agree_on_a_truncated_final_header() {
-    let spec = &BIG_COMPOSITE_HEADER;
-    let dir = fixture_dir(spec);
-    let schema = table_schema(spec);
-    let reader = open_reader(&dir).await;
-
-    for cut in [
-        HeaderCut::OneByte,
-        HeaderCut::InsideKey,
-        HeaderCut::InsideDeletionTime,
-    ] {
-        let (dec, hdr, keep) = truncation(&dir, cut);
-        // ONE buffer for all three: it begins at a partition header and is
-        // declared complete and unbounded, so every surface is being asked the
-        // same question about the same bytes.
-        let buf = &dec[hdr..keep];
-        let driver = driver_at_final_chunk(spec, &schema, &reader, buf);
-        let block = block_walk(spec, &schema, &reader, buf);
-        let meta = metadata_walk(spec, &schema, &reader, buf);
-
-        let verdicts = [
-            ("driver (at_final_chunk)", driver.is_err()),
-            ("block-emit walk", block.is_err()),
-            ("cell-metadata walk", meta.is_err()),
-        ];
-        assert!(
-            verdicts.iter().all(|(_, e)| *e) || verdicts.iter().all(|(_, e)| !*e),
-            "{cut:?}: the three walks DISAGREE about the same {} byte(s) — {}. A truncated \
-             header cannot be corruption on one surface and clean completion on another.",
-            buf.len(),
-            verdicts
-                .iter()
-                .map(|(n, e)| format!("{n}: {}", if *e { "Err" } else { "Ok" }))
-                .collect::<Vec<_>>()
-                .join(", ")
-        );
-        assert!(
-            verdicts.iter().all(|(_, e)| *e),
-            "{cut:?}: all three walks must REFUSE — they agreed on Ok, which reports a \
-             table with one fewer partition than the file holds"
-        );
-    }
-}
-
-/// C2 — a BOUNDED walk must still refuse a malformed INITIAL header.
-///
-/// The `bounded` flag disabled refusal for the WHOLE call, so
-/// `BufferExtent::Complete` plus `row_body_window: Some(..)` byte-resynced past
-/// the FIRST partition's header — before any window logic had run, and therefore
-/// before the window had introduced any uncertainty at all. The window's
-/// uncertainty begins only once the bounded row-body endpoint has been reached.
-///
-/// The resync WARN's absence is asserted alongside the refusal, because on this
-/// fixture the resync's garbage can trip #3782's row arm and produce an
-/// incidental `Err` — the same trap the BIG AC1 case documents. The positive
-/// control is the same bounded walk on a `Window` extent, which must still
-/// resync and log.
-#[tokio::test]
-async fn a_bounded_walk_still_refuses_a_malformed_initial_header() {
-    let spec = &BIG_COMPOSITE_HEADER;
-    let staged = fixture::stage_spec(spec, &fixture_dir(spec), "bounded-initial");
-    let schema = table_schema(spec);
-    let reader = open_reader(&staged.mutated_dir).await;
-    let mutated = fixture::stitched_data_section(&staged.mutated_dir);
-
-    // A well-formed clustering-slice request. The VALUES are immaterial — the
-    // malformed header sits at offset 0 and is reached before `row_body_window`
-    // is consulted at all — so they are taken from the PRISTINE section's first
-    // partition body so the request is a realistic one rather than arbitrary.
-    let pristine = fixture::stitched_data_section(&staged.control_dir);
-    let key_len = usize::from(u16::from_be_bytes([pristine[0], pristine[1]]));
-    let body_start = 2 + key_len + 12; // nb: fixed 12-byte partition DeletionTime
-    let window = (body_start, (body_start + 64).min(mutated.len()));
-
-    // POSITIVE CONTROL: on a Window the bounded walk must still take the
-    // tolerant resync and LOG it, which proves the sink captures and the needles
-    // are current.
-    let window_sink = LogSink::default();
-    let window_result = {
-        let _dispatch = tracing::subscriber::set_default(warn_subscriber(&window_sink));
-        bounded_walk(
-            spec,
-            &schema,
-            &reader,
-            &mutated,
-            window,
-            cqlite_core::storage::sstable::reader::BufferExtent::Window,
-        )
-    };
-    let window_logs = window_sink.text();
-    let observed: Vec<&str> = RESYNC_WARNS
-        .iter()
-        .copied()
-        .filter(|n| window_logs.contains(n))
-        .collect();
-    assert!(
-        !observed.is_empty(),
-        "positive control FAILED: a BOUNDED walk on a Window extent must still resync past \
-         the malformed initial header and LOG it. Expected one of {RESYNC_WARNS:?}; the walk \
-         answered {} and captured:\n{window_logs}",
-        window_result
-            .as_ref()
-            .map_or_else(|e| format!("Err({e})"), |k| format!("Ok({} keys)", k.len()))
-    );
-
-    // The subject: the SAME bounded walk on a COMPLETE buffer must refuse the
-    // initial header, and must not have resynced past it.
-    let sink = LogSink::default();
-    let got = {
-        let _dispatch = tracing::subscriber::set_default(warn_subscriber(&sink));
-        bounded_walk(
-            spec,
-            &schema,
-            &reader,
-            &mutated,
-            window,
-            cqlite_core::storage::sstable::reader::BufferExtent::Complete,
-        )
-    };
-    let logs = sink.text();
-    for needle in RESYNC_WARNS {
-        assert!(
-            !logs.contains(needle),
-            "C2: a bounded walk RESYNCHRONISED past the malformed INITIAL header — it logged \
-             {needle:?}. The row-body window introduces no uncertainty until its endpoint is \
-             reached, so partition 0's own header is as attributable as on an unbounded walk \
-             and must be REFUSED. (Capture proved live above: {observed:?}.) Captured:\n{logs}"
-        );
-    }
-    assert!(
-        got.is_err(),
-        "C2: a bounded walk over a PROVEN-COMPLETE buffer must refuse a malformed initial \
-         header; it answered Ok with {} key occurrence(s)",
-        got.map_or(0, |k| k.len())
-    );
-}
-
-/// The block-emit walk over `buf` declared a chunk-covering WINDOW — the
-/// tolerant extent, where a header may legitimately straddle the tail.
-///
-/// Used as the POSITIVE CONTROL for the resync-WARN absence assertions: it
-/// differs from `block_walk` in exactly one argument, so a resync it logs and
-/// `block_walk` does not is attributable to the extent alone.
-fn window_walk(
-    spec: &FixtureSpec,
-    schema: &TableSchema,
-    reader: &SSTableReader,
-    buf: &[u8],
-) -> Result<Vec<Vec<u8>>, String> {
-    use cqlite_core::storage::sstable::reader::BufferExtent;
-    let mut keys: Vec<Vec<u8>> = Vec::new();
-    framing_parser(spec)
-        .parse_block_emit_windowed(
-            buf,
-            BufferExtent::Window,
-            Some(schema),
-            reader,
-            None,
-            |(_t, k, _r)| {
-                keys.push(k.as_bytes().to_vec());
-                Ok(std::ops::ControlFlow::Continue(()))
-            },
-        )
-        .map_err(why)
-        .map(|()| keys)
-}
-
-/// The SLIDING driver MID-STREAM (`at_final_chunk == false`) — the legitimately
-/// tolerant leg, where a truncated header may still be completed by the next
-/// chunk and `NeedMore` is the correct answer.
-fn driver_mid_stream(
-    spec: &FixtureSpec,
-    schema: &TableSchema,
-    reader: &SSTableReader,
-    buf: &[u8],
-) -> Result<String, String> {
-    let mut emitted = 0usize;
-    framing_parser(spec)
-        .parse_one_partition_for_compaction(buf, Some(schema), reader, false, &mut |_row| {
-            emitted += 1;
-            Ok(std::ops::ControlFlow::Continue(()))
-        })
-        .map_err(why)
-        .map(|step| format!("{step:?} after emitting {emitted} row(s)"))
-}
-
-/// The block-emit walk over `buf` with a #954 row-body window — a BOUNDED walk,
-/// at the caller's stated extent.
-fn bounded_walk(
-    spec: &FixtureSpec,
-    schema: &TableSchema,
-    reader: &SSTableReader,
-    buf: &[u8],
-    window: (usize, usize),
-    extent: cqlite_core::storage::sstable::reader::BufferExtent,
-) -> Result<Vec<Vec<u8>>, String> {
-    let mut keys: Vec<Vec<u8>> = Vec::new();
-    framing_parser(spec)
-        .parse_block_emit_windowed(
-            buf,
-            extent,
-            Some(schema),
-            reader,
-            Some(window),
-            |(_t, k, _r)| {
-                keys.push(k.as_bytes().to_vec());
-                Ok(std::ops::ControlFlow::Continue(()))
-            },
-        )
-        .map_err(why)
-        .map(|()| keys)
 }
